BottomNavigationView:  - Small changes and fixes
