+++ app/src/main/java/com/volohov/currency/ui/predictionfragment/PredictionFragment.kt
@@ -1,10 +1,11 @@
-package com.volohov.currency
+package com.volohov.currency.ui.predictionfragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.volohov.currency.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ChartFragment.newInstance] factory method to
+ * Use the [PredictionFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ChartFragment : Fragment() {
+class PredictionFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class ChartFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_chart, container, false)
+        return inflater.inflate(R.layout.fragment_prediction, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class ChartFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ChartFragment.
+         * @return A new instance of fragment PredictionFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ChartFragment().apply {
+            PredictionFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
