+++ app/src/main/res/layout/fragment_currency.xml
@@ -5,7 +5,7 @@
     xmlns:card_view="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CurrencyFragment">
+    tools:context=".ui.currencyfragment.CurrencyFragment">
 
     <TextView
         android:id="@+id/last_date_update"
